Update to nextjs v9 and now v2
* nextjs v9 removes need for express and can use dynamic routing
* now v2 is all serverless
* updated all libraries
* fixed flickr images to be https to avoid mix-content issues
