+++ .babelrc
@@ -1,6 +1,5 @@
 {
     "presets": [
-      "next/babel",
-      "@zeit/next-typescript/babel"
+      "next/babel"
     ]
 }
