+++ components/FlickrImage.tsx
@@ -26,7 +26,7 @@ export class FlickrImage extends React.Component<FlickrImageProps> {
     } else if (size == 'k') {
       return this.props.url_k
     } else {
-      return `http://farm${this.props.farm}.staticflickr.com/${this.props.server}/${this.props.item_key}_${this.props.secret}_${size}.jpg`
+      return `https://farm${this.props.farm}.staticflickr.com/${this.props.server}/${this.props.item_key}_${this.props.secret}_${size}.jpg`
     }
   }
