+++ next.config.js
@@ -1,4 +1,3 @@
 // next.config.js
-const withTypescript = require('@zeit/next-typescript')
 const withCSS = require('@zeit/next-css')
-module.exports = withCSS(withTypescript())
+module.exports = withCSS()
