--- now.json
@@ -1,5 +1,0 @@
-{
-  "version": 1,
-  "name": "digest-photos",
-  "type": "docker"
-}
