+++ package.json
@@ -1,31 +1,28 @@
 {
   "scripts": {
-    "dev": "nodemon server/index.ts",
-    "build": "next build && tsc --project tsconfig.server.json",
-    "start": "NODE_ENV=production node .next/production-server/index.js"
+    "dev": "next",
+    "build": "next build",
+    "start": "next start"
   },
   "dependencies": {
     "@zeit/next-css": "^1.0.1",
-    "@zeit/next-typescript": "^1.1.1",
-    "express": "^4.16.4",
-    "flickr-sdk": "^3.7.0",
-    "mongodb": "^3.1.10",
-    "mongodb-stitch-browser-sdk": "^4.0.15-beta-1.0",
-    "mongodb-stitch-server-sdk": "^4.0.15-beta-1.0",
-    "next": "^7.0.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "semantic-ui-css": "^2.4.0",
-    "semantic-ui-react": "^0.82.5",
-    "typescript": "^3.1.6"
+    "flickr-sdk": "^3.9.0",
+    "mongodb": "^3.3.0",
+    "mongodb-stitch-browser-sdk": "^4.5.0",
+    "mongodb-stitch-server-sdk": "^4.5.0",
+    "next": "^9.0.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "semantic-ui-css": "^2.4.1",
+    "semantic-ui-react": "^0.87.3"
   },
   "devDependencies": {
-    "@types/mongodb": "^3.1.15",
-    "@types/express": "^4.16.0",
-    "@types/next": "^7.0.1",
-    "@types/node": "^10.12.10",
-    "@types/react": "^16.4.16",
-    "nodemon": "^1.18.6",
-    "ts-node": "^7.0.1"
+    "@types/node": "^12.7.2",
+    "@types/react": "^16.9.2",
+    "@types/react-dom": "^16.9.0",
+    "typescript": "^3.5.3"
+  },
+  "browser": {
+    "mongodb-stitch-server-sdk": false
   }
 }
