--- tsconfig.server.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": ".next/production-server"
-  },
-  "include": [
-    "server/**/*.ts", "operations/**/*.ts"
-  ]
-}
